// prol_educa_api/src/institutions/dto/create-institution.dto.ts
export class CreateInstitutionDto {
  
  trade_name: string;
    street: string;
    number: string;
    complement: string;
    district: string;
    city: string;
    state: string;
    postal_code: string;
    manager_name: string;
    manager_phone: string;
    manager_email: string;
    type: string;
    image_url: string;
    password: string;
    is_active: boolean;
    courses: any;
}